1. Changed the field type at CoffeeRecord. 2. Improved the code style. 3. Corrected the tests.

+++ app/G91/gonnykh.alexander/lab4/src/main/java/ru/vsu/amm/java/records/CoffeeRecord.java
@@ -4,8 +4,8 @@ import ru.vsu.amm.java.enums.CoffeeType;
 
 import java.time.LocalDateTime;
 
-public record CoffeeRecord(CoffeeType name, LocalDateTime date, long preparationTime, float cost) {
-    public float getPriceToTimeRatio() {
+public record CoffeeRecord(CoffeeType name, LocalDateTime date, long preparationTime, double cost) {
+    public double getPriceToTimeRatio() {
         if (preparationTime == 0) {
             throw new IllegalArgumentException("Preparation time cannot be zero.");
         }
